backend/requirements.txt
# WebSocket + Server
Flask==3.1.0
flask-socketio
eventlet
uuid

# Audio Transcription
openai-whisper
pyannote.audio
pydub

# Environment variables
python-dotenv

# Pinecone
pinecone-client

# LLM
groq
tf-keras
